configure: use the pkg.m4 macro library to handle pkg-config search
PKG_CHECK_MODULES is battle tested and provides some convenient
functionality such as standardizing on $PKG_CONFIG instead of
$PKGCONFIG, marking various additional variables as precious, and
upating the `./configure --help` text to notify you which variables can
be used to control the configure checks.

It's also shorter and simpler from the perspective of configure.ac
(and the added complexity in pkg.m4 is someone else's maintenance burden).

Note: pkg.m4 is GPL-2-or-later licensed. The GPL 2 is incompatible with
pdfio's Apache license. There are two reasons why this is still okay:

- The "or-later" clause means we can include pkg.m4 under the terms of
  the GPL-3, if we wish. This is the same license that autoconf itself
  uses, and is compatible with the Apache license.

- Like autoconf itself, pkg.m4 has a special clause permitting its
  distribution, when included "as part of a program that contains a
  configuration script generated by Autoconf", under whatever terms the
  rest of the program uses. Since that is what we are doing here, this
  means that pkg.m4 is licensed under the Apache license as long as it
  is used as part of the pdfio ./configure build system.